backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
drf-spectacular>=0.27.0
python-decouple>=3.8
django-cors-headers>=4.3.1
psutil>=7.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Environment Management
python-dotenv>=1.0.0

# Optional: Advanced connection pooling (for high-traffic production)
# pgbouncer-py>=1.0.0